Split logging of POST body from logging of URL, add body length
I was running into the problem that the post body was too big, DD has a
limit of 3MB, above which the request will be rejected. Therefore, in
order to easily monitor the current request size, I added the length of
the request body string.

Additionally, a very long request body led to log messages not being
recorded by the ELK stack I use. So in order for having the request
sizes logged even when a very long request body leads to broken logging,
I split the body from the rest and logged them separately.